requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
alembic=1.14.0=pyhd8ed1ab_1
anyio=4.6.2.post1=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310ha75aee5_5
arrow=1.3.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pyhd8ed1ab_0
attrs=24.2.0=pyh71513ae_0
babel=2.16.0=pyhd8ed1ab_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pyha770c72_0
bleach=6.2.0=pyhd8ed1ab_0
brotli-python=1.1.0=py310hf71b8c6_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.12.14=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
colorlog=6.9.0=pyh707e725_1
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dm-tree=0.1.8=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
enum34=1.1.10=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pyhd8ed1ab_0
future=1.0.0=pypi_0
gast=0.5.4=pypi_0
google-auth=2.29.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=3.1.1=py310hf71b8c6_1
grpcio=1.62.1=pypi_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
h5py=3.10.0=pypi_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.7=pyh29332c3_1
httpx=0.28.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.6=pypi_0
imbalanced-learn=0.12.2=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib_resources=6.4.5=pyhd8ed1ab_0
iniconfig=2.0.0=pypi_0
ipykernel=6.29.3=pypi_0
ipython=8.8.0=pypi_0
ipython_genutils=0.2.0=pyhd8ed1ab_1
ipywidgets=8.1.5=pyhd8ed1ab_0
iso8601=2.1.0=pypi_0
isoduration=20.11.0=pyhd8ed1ab_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
json5=0.10.0=pyhd8ed1ab_0
jsonpointer=3.0.0=py310hff52083_1
jsonschema=4.23.0=pyhd8ed1ab_0
jsonschema-specifications=2024.10.1=pyhd8ed1ab_0
jsonschema-with-format-nongpl=4.23.0=hd8ed1ab_0
jupyter=1.1.1=pyhd8ed1ab_0
jupyter-client=8.6.1=pypi_0
jupyter-lsp=2.2.5=pyhd8ed1ab_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_console=6.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
jupyter_events=0.10.0=pyhd8ed1ab_0
jupyter_server=2.14.2=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_0
jupyterlab=4.3.1=pyhff2d567_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
jupyterlab_server=2.27.3=pyhd8ed1ab_0
jupyterlab_widgets=3.0.13=pyhd8ed1ab_0
keras=2.15.0=pypi_0
keras-tcn=3.5.0=pypi_0
keras-tuner=1.0.3=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.3=h659f571_0
kt-legacy=1.0.5=pypi_0
ld_impl_linux-64=2.43=h712a8e2_2
libblas=3.9.0=26_linux64_openblas
libcblas=3.9.0=26_linux64_openblas
libclang=18.1.1=pypi_0
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libgomp=14.2.0=h77fa898_1
liblapack=3.9.0=26_linux64_openblas
libnsl=2.0.1=hd590300_0
libopenblas=0.3.28=pthreads_h94d23a6_1
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.47.0=hadc24fc_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
mako=1.3.8=pyhd8ed1ab_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pyhd8ed1ab_0
ml-dtypes=0.2.0=pypi_0
namex=0.0.7=pypi_0
nbclassic=1.1.0=pyhd8ed1ab_0
nbclient=0.10.1=pyhd8ed1ab_0
nbconvert-core=7.16.4=pyhd8ed1ab_1
nbformat=5.10.4=pyhd8ed1ab_0
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
netron=8.0.5=pypi_0
networkx=3.2.1=pypi_0
notebook=6.5.7=pyha770c72_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=1.26.4=py310hb13e2d6_0
nvidia-cublas-cu12=12.3.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.3.101=pypi_0
nvidia-cuda-nvcc-cu12=12.3.107=pypi_0
nvidia-cuda-nvrtc-cu12=12.3.107=pypi_0
nvidia-cuda-runtime-cu12=12.3.101=pypi_0
nvidia-cudnn-cu12=8.9.7.29=pypi_0
nvidia-cufft-cu12=11.0.12.1=pypi_0
nvidia-curand-cu12=10.3.4.107=pypi_0
nvidia-cusolver-cu12=11.5.4.101=pypi_0
nvidia-cusparse-cu12=12.2.0.103=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.4.0=hb9d3cd8_0
opt-einsum=3.3.0=pypi_0
optree=0.11.0=pypi_0
optuna=4.1.0=pyhd8ed1ab_1
overrides=7.7.0=pyhd8ed1ab_0
packaging=24.0=pypi_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parse=1.6.5=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.3.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.2.0=pypi_0
pluggy=1.4.0=pypi_0
prometheus_client=0.21.0=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pypi_0
prompt_toolkit=3.0.48=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure-eval=0.2.2=pypi_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pyhd8ed1ab_0
pygments=2.17.2=pypi_0
pyparser=1.0=pypi_0
pyparsing=3.1.2=pypi_0
pyserial=3.5=pypi_0
pysocks=1.7.1=pyha2e5f31_6
pytest=8.1.1=pypi_0
python=3.10.14=hd12c33a_0_cpython
python-dateutil=2.9.0=pypi_0
python-fastjsonschema=2.21.0=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python_abi=3.10=5_cp310
pytz=2024.1=pypi_0
pyyaml=6.0.2=py310ha75aee5_1
pyzmq=25.1.2=pypi_0
qkeras=0.9.0=pypi_0
readline=8.2=h8228510_1
referencing=0.35.1=pyhd8ed1ab_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rich=13.7.1=pypi_0
rpds-py=0.21.0=py310h505e2c1_0
rsa=4.9=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
send2trash=1.8.3=pyh0d859eb_0
setuptools=75.6.0=pyhff2d567_1
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
sqlalchemy=2.0.36=py310ha75aee5_0
stack_data=0.6.2=pyhd8ed1ab_0
tensorboard=2.15.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0=pypi_0
tensorflow-addons=0.23.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
tensorflow-model-optimization=0.8.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pyh0d859eb_0
tf-keras=2.15.0=pypi_0
threadpoolctl=3.4.0=pypi_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.2.1=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
typeguard=2.13.3=pypi_0
types-python-dateutil=2.9.0.20241003=pyhff2d567_0
typing-extensions=4.10.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
typing_utils=0.1.0=pyhd8ed1ab_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pyhd8ed1ab_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=24.8.0=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.8.0=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.45.1=pyhd8ed1ab_0
widgetsnbextension=4.0.13=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.5=h3b0a872_7
zipp=3.17.0=pypi_0
zstandard=0.23.0=py310ha39cb0e_1
zstd=1.5.6=ha6fb4c9_0
